Split type impl generic template tests
